Implement seq update

+++ packages-rs/drainpipe/src/schema.rs
@@ -3,7 +3,7 @@
 diesel::table! {
     dead_letter_queue (rowid) {
         rowid -> Integer,
-        seq -> Integer,
+        seq -> BigInt,
         msg -> Text,
     }
 }
@@ -11,8 +11,11 @@ diesel::table! {
 diesel::table! {
     drainpipe (rowid) {
         rowid -> Integer,
-        seq -> Integer,
+        seq -> BigInt,
     }
 }
 
-diesel::allow_tables_to_appear_in_same_query!(dead_letter_queue, drainpipe,);
+diesel::allow_tables_to_appear_in_same_query!(
+    dead_letter_queue,
+    drainpipe,
+);
